requirements.txt
aiohttp==1.0.5
aiosocks==0.1.7
async-timeout==1.1.0
beautifulsoup4==4.5.1
bs4==0.0.1
cairocffi==0.7.2
CairoSVG==2.0.0
cffi==1.9.1
chardet==2.3.0
ChatterBot==0.5.3
cleverbot==1.1.1
click==6.6
cryptography==1.7.1
cssselect==1.0.0
cycler==0.10.0
discord.py==0.16.0
Flask==0.12
future==0.16.0
idna==2.2
imgurpython==1.1.7
inflect==0.2.5
itsdangerous==0.24
jaraco.itertools==2.0
Jinja2==2.8
jpglitch==0.2
jsondatabase==0.1.6
jsonpickle==0.9.3
lxml==3.7.1
MarkupSafe==0.23
matplotlib==1.5.3
more-itertools==2.4.1
multidict==2.1.4
nltk==3.2.1
numpy==1.11.3
Pillow==3.4.2
pyasn1==0.1.9
pycparser==2.17
pyfiglet==0.7.5
pymongo==3.4.0
PyMySQL==0.7.9
pyparsing==2.1.10
python-aalib==0.3.2
python-dateutil==2.6.0
python-magic==0.4.12
python-twitter==3.2
pytz==2016.10
requests==2.12.4
requests-oauthlib==0.7.0
six==1.10.0
SQLAlchemy==1.1.5.dev0
steamapi==0.1
tabulate==0.7.7
textblob==0.11.1
texttable==0.8.7
timestring==1.6.2
tinycss==0.4
urbandict==0.5
vdf==2.1
Wand==0.4.4
websockets==3.2
Werkzeug==0.11.11
wolframalpha==3.0
wordcloud==1.2.1
xmltodict==0.10.2
